+++ src/js/JSONEditor.js
@@ -184,7 +184,8 @@ JSONEditor.VALID_OPTIONS = [
   'escapeUnicode', 'history', 'search', 'mode', 'modes', 'name', 'indentation',
   'sortObjectKeys', 'navigationBar', 'statusBar', 'mainMenuBar', 'languages', 'language', 'enableSort', 'enableTransform',
   'maxVisibleChilds', 'onValidationError',
-  'modalAnchor', 'popupAnchor'
+  'modalAnchor', 'popupAnchor',
+  'createQuery', 'executeQuery', 'queryDescription'
 ]
 
 /**